Create Rock_Paper_Scissors_Game.py
Pseudocode: Game rock, paper, scissors. 
Ask which one to choose. 0 to rock, 1 to paper, 2 to scissors. 
Print choice.
If por printing the choice. 
Print computer random choice.
Make a list of possibilities then use random function to print one of them. 
Print who win. 
Make a if loop to describe in which conditions the person lose or win.